requirements.txt
Flask==2.3.3
pytesseract==0.3.10
PyMuPDF==1.22.5
Pillow==10.0.0
schedule==1.2.0
tabulate==0.9.0
google-api-python-client==2.88.0
google-auth==2.22.0
gspread==6.9.0
oauth2client==4.1.3
opencv-python-headless==4.8.1.78
numpy==1.26.0
